Squashed commit of the following:
commit 5813887ccf93be0d33a408c72e45e8e63092a680
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Wed Jun 7 22:43:54 2023 +0700

    fixes

commit 5362ce9f2b8c82906ee1ab0f6d9f36fee4f6fa9d
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Thu Jun 1 01:44:24 2023 +0700

    MyBodyScreen logic completed

commit 778166149e2d449fcaa8237fd8dd8c232d6a89d6
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Mon May 29 20:04:04 2023 +0700

    MyBodyScreen layout completed

commit 5409771089c264327a3dd885a5b246bec4d270b6
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Sat May 27 22:43:48 2023 +0700

    2 screens + ctrl+alt+l

commit 5aeffb5b6096ca30438ecc111b0ba52c3ed000c4
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Fri May 26 01:49:06 2023 +0700

    3 Exercises screens: Crunch, Squats, Plank

commit f0cdb329f7711df17a87dd304ff24c09b2e16a15
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Wed May 24 20:21:14 2023 +0700

    Exercise screen layout, Event and state

commit b964f66e8a4d107a100184329729fe499f3fecbb
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Fri May 19 16:40:26 2023 +0700

    UseCase usages update

commit f210c40d6e34194c158281f965af850adb39edc9
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Fri May 19 16:38:49 2023 +0700

    Small updates in navigation and network

commit f0b4f16c36b274ef520745bc400a2b4fcf2e4702
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Fri May 19 16:37:39 2023 +0700

    Added room DB for trainings

commit 8b750fde98a24dc39c7d337b4d88376828b8a729
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Fri May 19 16:36:05 2023 +0700

    Added storages for user email and first run

commit e20bdd1773489e2c06ae7656636b11a1e6ab2820
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Tue May 16 19:17:59 2023 +0700

    Added extensions for Modifier

    With this extensions you could fill all avaliable space in infinite scopes (in column). It helps to show progress bar at the center of white column on MainScreen

commit 477d914e6512786590e7a8d462b1ca75254f6496
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Tue May 16 18:51:22 2023 +0700

    Added token saving

    Added storage for token, usecases for token, updated interceptor for doing requests with authorization header

commit e2ce1828195296cc4623a6476dc59ec62a26ef33
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Tue May 16 14:31:58 2023 +0700

    Some layout fixes

commit 94d3b8e4d4441479d2c0b6ba6d93caa635fb88df
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Tue May 16 14:19:10 2023 +0700

    DI fixes + errors with alert dialog

commit 9837a625799ad7e9e29c34b1943bf0ad2e51fc54
Merge: e3aa69f 5d3eac8
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Sun May 14 22:11:53 2023 +0700

    Merge remote-tracking branch 'origin/screens/main' into screens/main

commit e3aa69fee03fdee93ac70bdcd340e121d66aced0
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Sun May 14 22:11:16 2023 +0700

    Added state + event for main screen

commit 5d3eac84655b0dd60577dc737abb45b081d0c9ea
Author: Velikanova Sofia <solntselikaya.sofia@yandex.ru>
Date:   Sun May 14 22:11:16 2023 +0700

    Added state + event for main screen

+++ app/src/main/java/com/example/superfit/data/dto/AccessTokenDto.kt
@@ -4,7 +4,7 @@ import com.example.superfit.domain.model.AccessTokenModel
 
 data class AccessTokenDto(
     val access_token: String,
-    val expired: Int
+    val expired: Long
 )
 
 fun AccessTokenDto.toAccessTokenModel(): AccessTokenModel =
